playground/mac_agents/requirements.txt
google-generativeai>=0.7.0
requests>=2.31.0
pytest>=8.0.0
